requirements.txt
# Django Core
Django>=5.2.0,<6.0.0

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0

# WebSocket Support (Django Channels)
channels>=4.0.0
daphne>=4.0.0
# Optional for production with Redis:
# channels-redis>=4.0.0

# CORS Headers
django-cors-headers>=4.0.0

# AWS S3 Storage
django-storages>=1.14.0
boto3>=1.26.0

# Image Processing (required for ImageField)
Pillow>=10.0.0

# Environment Variables
python-dotenv>=1.0.0

# WebSocket Testing
websockets>=12.0

# Firebase Admin SDK
firebase-admin>=7.0.0
